delete confirmation dialog

+++ src/app/modules/home/exams/exams.component.ts
@@ -5,6 +5,8 @@ import { TaskService } from 'src/app/core/services/db-utils/task.service';
 import { ContestantService } from 'src/app/core/services/db-utils/contestant.service';
 import { SubmissionService } from 'src/app/core/services/db-utils/submission.service';
 import * as _ from 'lodash';
+import { MatDialog, MatDialogConfig } from '@angular/material';
+import { ConfirmDialogComponent } from 'src/app/core/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-exams',
@@ -16,6 +18,7 @@ export class ExamsComponent implements OnInit {
   private examList: IExam[] = [];
 
   constructor(
+    private dialog: MatDialog,
     private examService: ExamService,
     private taskService: TaskService,
     private contestantService: ContestantService,
@@ -29,10 +32,22 @@ export class ExamsComponent implements OnInit {
   }
 
   remove(examId: number) {
-    this.examService.remove(examId);
-    this.taskService.removeByExamId(examId);
-    this.contestantService.removeByExamId(examId);
-    this.submissionService.removeByExamId(examId);
-    _.remove(this.examList, (exam: IExam) => exam.examId == examId);
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.data = {
+      title: 'Delete Exam Confirmation',
+      message: `Are you sure you want to delete exam ${examId}?`
+    };
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe((res: boolean) => {
+      if (res) {
+        this.examService.remove(examId);
+        this.taskService.removeByExamId(examId);
+        this.contestantService.removeByExamId(examId);
+        this.submissionService.removeByExamId(examId);
+        _.remove(this.examList, (exam: IExam) => exam.examId == examId);
+      }
+    });
   }
 }
